applications/python/service-a/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
prometheus-client
requests
loguru